Hierarchial Clustering (#52)
* UCB(Reinforcement Learning)

Upper Confidence Bound Learning.The Given Code is used to find which out of the 10 ads to be displayed on website for maximum Click Through Response by the user. 
The dataset is virtual showing what the ith user would have done if one of the 10 ads was shown to him i.e. 1 specifying he would have clicked it and 0 means he would have ignored the add

* Hierarchial Clustering

Hierarchial Clustering is type of clustering technique in which initially all datapoints are treated as seperate clusters.Then two clusters are merged with each together to for N-1 clusters (initially there are N datapoints).This process until a single cluster is formed.Dendrogram plot is formed to see which clusters were merged during the process and by studying the "Dendrograms" we decide the number of clusters to be formed
To form the clusters I have used the "Ward" method which works on the principle that the within cluster 'variance' is minimised after two clusters are merged.
To get a brief overview I have provided a link..http://www.stat.cmu.edu/~ryantibs/datamining/lectures/05-clus2-marked.pdf