Association Rule Learning (Apriori Method) (#55)
* UCB(Reinforcement Learning)

Upper Confidence Bound Learning.The Given Code is used to find which out of the 10 ads to be displayed on website for maximum Click Through Response by the user. 
The dataset is virtual showing what the ith user would have done if one of the 10 ads was shown to him i.e. 1 specifying he would have clicked it and 0 means he would have ignored the add

* Hierarchial Clustering

Hierarchial Clustering is type of clustering technique in which initially all datapoints are treated as seperate clusters.Then two clusters are merged with each together to for N-1 clusters (initially there are N datapoints).This process until a single cluster is formed.Dendrogram plot is formed to see which clusters were merged during the process and by studying the "Dendrograms" we decide the number of clusters to be formed
To form the clusters I have used the "Ward" method which works on the principle that the within cluster 'variance' is minimised after two clusters are merged.
To get a brief overview I have provided a link..http://www.stat.cmu.edu/~ryantibs/datamining/lectures/05-clus2-marked.pdf

* Apriori Algorithm(Association Rule Learning)

Apriori Algorithm is a part of Association Rule Learning where associations between variables or objects in the real world is found out.One of the biggest applications of this algorithm is video recommendations.While using You Tube we see that on the side recommended videos section is there.It is build using data from user as to which videos people watched after watching the specific video.
In my model I have found association between various goods sold at a shopping mart(like:Walmart).In this dataset contains list of 7500 customers and their shopping list.By fitting the apriori model in this dataset I have found the association between these goods i.e.the goods that people prefer to buy together.By getting this data the store can increase their profit by placing the products forming a strong association close by.
The function is imported from a python class named under apyori.py. The final association table can be printed and seen on the IDE itself.The various associations are arranged in descending order of Lift.
Intuition and important are provided in the pictures.
            ''The rule is that Association between two or more items is strong when the Lift is high.''